consolidate jdk11 and jdk14 tests, cleanup imports and debugging output
