+++ pom.xml
@@ -200,27 +200,36 @@
 				</plugins>
 			</build>
 		</profile>
-		<!-- Exclude any tests that require JDK 11+ from test run -->
+		<!-- Include any tests that require JDK11+ -->
 		<profile>
-			<id>until-jdk11</id>
+			<id>jdk11ge</id>
 			<activation>
-				<!-- Use exclusive 11 range instead of inclusive 10, because an upper bound ",10]" is likely not to include most releases of 10,
-				   since they will have an additional "patch" release such as _05 that is not taken into consideration in the above range.
-				   See also http://maven.apache.org/guides/introduction/introduction-to-profiles.html#Details_on_profile_activation -->
-				<jdk>[1.5,11)</jdk>
+				<jdk>[11,</jdk>
 			</activation>
 			<build>
 				<pluginManagement>
 					<plugins>
 						<plugin>
+							<groupId>org.codehaus.mojo</groupId>
+							<artifactId>build-helper-maven-plugin</artifactId>
+							<executions>
+								<execution>
+									<id>add-test-source</id>
+									<phase>generate-test-sources</phase>
+									<goals>
+										<goal>add-test-source</goal>
+									</goals>
+									<configuration>
+										<sources>
+											<source>test-jdk11</source>
+										</sources>
+									</configuration>
+								</execution>
+							</executions>
+						</plugin>
+						<plugin>
 							<groupId>org.apache.maven.plugins</groupId>
-							<artifactId>maven-surefire-plugin</artifactId>
-							<version>2.22.0</version>
-							<configuration>
-								<excludes>
-									<exclude>jdk11/**</exclude>
-								</excludes>
-							</configuration>
+							<artifactId>maven-compiler-plugin</artifactId>
 						</plugin>
 					</plugins>
 				</pluginManagement>
@@ -247,6 +256,7 @@
 									</goals>
 									<configuration>
 										<sources>
+											<source>test-jdk11</source>
 											<source>test-jdk14</source>
 										</sources>
 									</configuration>
